auth/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
Authlib==0.15.5
Babel==2.9.1
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
flasgger==0.9.5
Flask==2.0.3
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-OpenTracing==1.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.0
jaeger-client==4.8.0
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
mistune==2.0.2
opentracing==2.4.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.31
threadloop==1.0.2
thrift==0.15.0
tomli==2.0.1
tornado==6.1
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
